Minor changes based on a review by Mark Taylor.
Changing the description of utype to at least reference the major utype
users (the previous language pointing to VOTable was highly misleading).

BaseParam attributes no longer refer to "columns" but instead to
"parameters and columns".

Updated arraysize pattern.  It has overgenerated in 1.1; I think
what's in now matches the prose in VOTable (the VOTable schema doesn't
restrict it).